semi-working backend e2e test
