+++ test/kubernetes/e2e/features/backends/suite.go
@@ -1,0 +1,99 @@
+package backends
+
+import (
+	"context"
+	"net/http"
+	"path/filepath"
+
+	"github.com/onsi/gomega"
+	"github.com/stretchr/testify/suite"
+	appsv1 "k8s.io/api/apps/v1"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
+	"github.com/kgateway-dev/kgateway/v2/api/v1alpha1"
+	"github.com/kgateway-dev/kgateway/v2/pkg/utils/fsutils"
+	"github.com/kgateway-dev/kgateway/v2/pkg/utils/kubeutils"
+	"github.com/kgateway-dev/kgateway/v2/pkg/utils/kubeutils/kubectl"
+	"github.com/kgateway-dev/kgateway/v2/pkg/utils/requestutils/curl"
+	"github.com/kgateway-dev/kgateway/v2/test/gomega/matchers"
+	"github.com/kgateway-dev/kgateway/v2/test/kubernetes/e2e"
+)
+
+var _ e2e.NewSuiteFunc = NewTestingSuite
+
+var (
+	baseManifest = filepath.Join(fsutils.MustGetThisDir(), "inputs/base.yaml")
+	// separate manifest to allow creation independently of routing config
+	backendManifest = filepath.Join(fsutils.MustGetThisDir(), "inputs/backend.yaml")
+
+	proxyObjMeta = metav1.ObjectMeta{
+		Name:      "gw",
+		Namespace: "default",
+	}
+	proxyDeployment = &appsv1.Deployment{ObjectMeta: proxyObjMeta}
+	proxyService    = &corev1.Service{ObjectMeta: proxyObjMeta}
+
+	curlPodExecOpt = kubectl.PodExecOptions{
+		Name:      "curl",
+		Namespace: "curl",
+		Container: "curl",
+	}
+
+	expectedResponse = &matchers.HttpResponse{
+		StatusCode: http.StatusOK,
+		Body:       gomega.ContainSubstring("sunt aut facere repellat provident occaecati excepturi optio reprehenderit"),
+	}
+
+	backendMeta = metav1.ObjectMeta{
+		Name:      "json-upstream",
+		Namespace: "default",
+	}
+	be = v1alpha1.Backend{
+		ObjectMeta: backendMeta,
+	}
+)
+
+type testingSuite struct {
+	suite.Suite
+	ctx              context.Context
+	testInstallation *e2e.TestInstallation
+}
+
+func NewTestingSuite(ctx context.Context, testInst *e2e.TestInstallation) suite.TestingSuite {
+	return &testingSuite{
+		ctx:              ctx,
+		testInstallation: testInst,
+	}
+}
+
+func (s *testingSuite) TestConfigureBackingDestinationsWithUpstream() {
+	s.T().Cleanup(func() {
+		err := s.testInstallation.Actions.Kubectl().DeleteFile(s.ctx, baseManifest)
+		s.NoError(err, "can delete manifest")
+		err = s.testInstallation.Actions.Kubectl().DeleteFile(s.ctx, backendManifest)
+		s.NoError(err, "can delete manifest")
+		s.testInstallation.Assertions.EventuallyObjectsNotExist(s.ctx, proxyService, proxyDeployment)
+	})
+
+	err := s.testInstallation.Actions.Kubectl().ApplyFile(s.ctx, baseManifest)
+	s.Require().NoError(err, "can apply base manifest")
+
+	// apply the upstream manifest separately, after the route table is applied, to ensure it can be applied after the route table
+	err = s.testInstallation.Actions.Kubectl().ApplyFile(s.ctx, backendManifest)
+	s.Require().NoError(err, "can apply Backend manifest")
+
+	s.testInstallation.Assertions.EventuallyObjectsExist(s.ctx, proxyService, proxyDeployment)
+	s.testInstallation.Assertions.AssertEventualCurlResponse(
+		s.ctx,
+		curlPodExecOpt,
+		[]curl.Option{
+			curl.WithHost(kubeutils.ServiceFQDN(proxyService.ObjectMeta)),
+			curl.WithHostHeader("example.com"),
+			curl.WithPath("/posts/1"),
+		},
+		expectedResponse,
+	)
+
+	s.testInstallation.Assertions.EventuallyObjectsExist(s.ctx, &be)
+}
