--- test/kubernetes/e2e/features/upstreams/inputs/upstream-for-route.yaml
@@ -1,10 +1,0 @@
----
-apiVersion: gloo.solo.io/v1
-kind: Upstream
-metadata:
-  name: json-upstream
-spec:
-  static:
-    hosts:
-      - addr: jsonplaceholder.typicode.com
-        port: 80
