+++ test/kubernetes/e2e/tests/kgateway_tests.go
@@ -15,9 +15,9 @@ import (
 	// "github.com/kgateway-dev/kgateway/v2/test/kubernetes/e2e/features/port_routing"
 	// "github.com/kgateway-dev/kgateway/v2/test/kubernetes/e2e/features/route_delegation"
 	// "github.com/kgateway-dev/kgateway/v2/test/kubernetes/e2e/features/route_options"
+	"github.com/kgateway-dev/kgateway/v2/test/kubernetes/e2e/features/backends"
 	"github.com/kgateway-dev/kgateway/v2/test/kubernetes/e2e/features/services/httproute"
 	"github.com/kgateway-dev/kgateway/v2/test/kubernetes/e2e/features/services/tcproute"
-	// "github.com/kgateway-dev/kgateway/v2/test/kubernetes/e2e/features/upstreams"
 	// "github.com/kgateway-dev/kgateway/v2/test/kubernetes/e2e/features/virtualhost_options"
 )
 
@@ -31,7 +31,7 @@ func KubeGatewaySuiteRunner() e2e.SuiteRunner {
 	// kubeGatewaySuiteRunner.Register("ListenerOptions", listener_options.NewTestingSuite)
 	// kubeGatewaySuiteRunner.Register("RouteOptions", route_options.NewTestingSuite)
 	// kubeGatewaySuiteRunner.Register("VirtualHostOptions", virtualhost_options.NewTestingSuite)
-	// kubeGatewaySuiteRunner.Register("Upstreams", upstreams.NewTestingSuite)
+	kubeGatewaySuiteRunner.Register("Backends", backends.NewTestingSuite)
 	kubeGatewaySuiteRunner.Register("HTTPRouteServices", httproute.NewTestingSuite)
 	kubeGatewaySuiteRunner.Register("TCPRouteServices", tcproute.NewTestingSuite)
 	// kubeGatewaySuiteRunner.Register("HeadlessSvc", headless_svc.NewK8sGatewayHeadlessSvcSuite)
